requirements.txt
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pillow==11.0.0
python-dotenv==1.0.1
python-telegram-bot==20.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.13.3
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.33.0
typing_extensions==4.12.2
urllib3==2.2.3
